requirements.txt
Flask==2.2.3
numpy==1.24.1
opencv_python==4.7.0.68
Pillow==9.5.0
tensorflow==2.12.0
